requirements.txt
# Need to add google module

pbr>=0.11,<2.0
alembic>=0.7.2
Babel>=1.3
tox
jsonschema
testtools>=0.9.36,!=1.2.0
PyYAML
pygoogle-simple
